+++ koupleless-adapter-springboot-2.3/pom.xml
@@ -35,7 +35,7 @@
 
         <dependency>
             <groupId>com.alipay.sofa.koupleless</groupId>
-            <artifactId>koupleless-common</artifactId>
+            <artifactId>koupleless-base-plugin</artifactId>
             <version>${revision}</version>
             <scope>provided</scope>
         </dependency>
