+++ koupleless-adapter-springboot-2.4.6/pom.xml
@@ -36,7 +36,7 @@
 
         <dependency>
             <groupId>com.alipay.sofa.koupleless</groupId>
-            <artifactId>koupleless-common</artifactId>
+            <artifactId>koupleless-base-plugin</artifactId>
             <version>${revision}</version>
             <scope>provided</scope>
         </dependency>
